chore: upgrade/downgrade some pkgs to fix warning
