+++ package.docker.json
@@ -13,11 +13,13 @@
     "@editorjs/checklist": "^1.1.0",
     "@editorjs/delimiter": "^1.1.0",
     "@editorjs/editorjs": "^2.15.0",
+    "@editorjs/header": "^2.3.0",
     "@editorjs/inline-code": "^1.3.0",
     "@editorjs/list": "^1.4.0",
     "@editorjs/marker": "^1.2.1",
-    "@groupher/editor-quote": "^2.3.0",
     "@groupher/editor-header": "^2.3.0",
+    "@groupher/editor-quote": "^2.3.0",
+    "@next/bundle-analyzer": "^9.0.0",
     "@sentry/browser": "^4.6.4",
     "@sentry/node": "^4.6.4",
     "@tippy.js/react": "^2.2.1",
@@ -25,8 +27,8 @@
     "@zeit/next-source-maps": "0.0.4-canary.1",
     "accepts": "^1.3.4",
     "antd": "3.8.4",
-    "apollo-cache-inmemory": "1.6.2",
-    "apollo-client": "2.6.3",
+    "apollo-cache-inmemory": "1.6.5",
+    "apollo-client": "2.6.6",
     "apollo-link": "1.2.5",
     "apollo-link-error": "^1.0.9",
     "apollo-link-http": "1.5.14",
@@ -42,7 +44,7 @@
     "draft-js-plugins-editor": "2.1.1",
     "express": "^4.16.4",
     "glob": "^7.1.2",
-    "graphql": "14.4.0",
+    "graphql": "14.5.5",
     "graphql-request": "^1.6.0",
     "graphql-tag": "^2.6.1",
     "graphql-voyager": "^1.0.0-rc.26",
@@ -58,10 +60,11 @@
     "mobx-react-lite": "1.4.1",
     "mobx-state-tree": "3.14.1",
     "module-alias": "^2.0.1",
-    "next": "9.0.2",
+    "next": "9.1.6",
     "next-compose-plugins": "^2.2.0",
-    "next-offline": "^4.0.2",
-    "next-seo": "1.11.1",
+    "next-offline": "4.0.6",
+    "next-progressbar": "^1.0.0",
+    "next-seo": "3.2.0",
     "overlayscrollbars": "1.9.1",
     "overlayscrollbars-react": "^0.1.0",
     "path-match": "^1.2.4",
@@ -97,8 +100,8 @@
     "shortid": "^2.2.8",
     "store": "^2.0.12",
     "stringz": "^2.0.0",
-    "styled-components": "4.1.3",
-    "timeago-react": "2.0.0",
+    "styled-components": "4.3.2",
+    "timeago-react": "3.0.0",
     "uuid": "^3.3.2"
   },
   "devDependencies": {
