+++ package.json
@@ -23,8 +23,8 @@
     "clean": "shjs ./utils/scripts/clean.js",
     "gen": "plop --plopfile ./utils/scripts/generators/index.js",
     "analyze": "cross-env ANALYZE=true next build",
-    "lint": "eslint --ext .js --fix pages/ components/ containers/ stores/ utils/ services/",
-    "lint:staged": "git diff --cached --name-only --diff-filter=ACRM | grep \".js$\" | xargs eslint --ext .js",
+    "lint": "./node_modules/.bin/eslint --ext .js --fix src/ utils/",
+    "lint:staged": "git diff --cached --name-only --diff-filter=ACRM | grep \".js$\" | xargs ./node_modules/.bin/eslint --ext .js",
     "test": "cross-env NODE_ENV=test jest --config .jest.config.js",
     "test:dev": "cross-env NODE_ENV=test jest --config .jest.config.js",
     "test:watch": "npm run test -- --watch",
@@ -90,11 +90,11 @@
     "mobx-react-lite": "1.4.1",
     "mobx-state-tree": "3.14.1",
     "module-alias": "^2.0.1",
-    "next": "9.0.6",
+    "next": "9.1.6",
     "next-compose-plugins": "^2.2.0",
-    "next-offline": "^4.0.2",
+    "next-offline": "4.0.6",
     "next-progressbar": "^1.0.0",
-    "next-seo": "1.11.1",
+    "next-seo": "3.2.0",
     "overlayscrollbars": "1.9.1",
     "overlayscrollbars-react": "^0.1.0",
     "path-match": "^1.2.4",
@@ -104,10 +104,10 @@
     "pubsub-js": "^1.6.0",
     "ramda": "0.25.0",
     "randomcolor": "^0.5.3",
-    "react": "16.10.0",
+    "react": "16.9.0",
     "react-calendar-heatmap": "1.8.1",
     "react-content-loader": "3.4.2",
-    "react-dom": "16.10.0",
+    "react-dom": "16.9.0",
     "react-helmet": "^5.2.1",
     "react-highlight-words": "^0.16.0",
     "react-intl": "3.0.0",
@@ -131,7 +131,7 @@
     "store": "^2.0.12",
     "stringz": "^2.0.0",
     "styled-components": "4.3.2",
-    "timeago-react": "2.0.0",
+    "timeago-react": "3.0.0",
     "uuid": "^3.3.2"
   },
   "devDependencies": {
