+++ src/containers/CheatsheetThread/parser.js
@@ -4,16 +4,8 @@ const groupSpliter = '{{ ::group:: }}'
 const cardsHeaderSpliter = '{{ ::cards-header:: }}'
 const cardItemSpliter = '{{ ::card-item:: }}'
 
-const getCardHeader = R.compose(
-  R.trim,
-  R.head,
-  R.split(cardsHeaderSpliter)
-)
-const getCardList = R.compose(
-  R.trim,
-  R.nth(1),
-  R.split(cardsHeaderSpliter)
-)
+const getCardHeader = R.compose(R.trim, R.head, R.split(cardsHeaderSpliter))
+const getCardList = R.compose(R.trim, R.nth(1), R.split(cardsHeaderSpliter))
 
 const getCardItems = R.compose(
   R.map(R.trim),
@@ -25,10 +17,6 @@ const formatFromer = v => ({
   cards: getCardItems(v),
 })
 
-const parser = R.compose(
-  R.map(formatFromer),
-  R.split(groupSpliter),
-  R.trim
-)
+const parser = R.compose(R.map(formatFromer), R.split(groupSpliter), R.trim)
 
 export default parser
