+++ src/containers/PostsThread/logic.js
@@ -250,7 +250,6 @@ const ErrSolver = [
 export const useInit = _store =>
   useEffect(() => {
     store = _store
-    // log('effect init')
     sub$ = sr71$.data().subscribe($solver(DataSolver, ErrSolver))
     /*
          NOTE: city communities list is not supported by SSR
