+++ src/pages/job.js
@@ -3,19 +3,21 @@ import { Provider } from 'mobx-react'
 import R from 'ramda'
 import { BlogJsonLd } from 'next-seo'
 
+// eslint-disable-next-line import/named
 import { PAGE_SIZE, SITE_URL } from '@config'
 import { TYPE, ROUTE, THREAD } from '@constant'
 import {
   getJwtToken,
-  nilOrEmpty,
   makeGQClient,
   parseURL,
+  nilOrEmpty,
   ssrAmbulance,
   parseTheme,
 } from '@utils'
 import initRootStore from '@stores/init'
 
 import AnalysisService from '@services/Analysis'
+
 import GlobalLayout from '@containers/GlobalLayout'
 import ThemeWrapper from '@containers/ThemeWrapper'
 import MultiLanguage from '@containers/MultiLanguage'
@@ -24,33 +26,33 @@ import Doraemon from '@containers/Doraemon'
 import Route from '@containers/Route'
 import Header from '@containers/Header'
 import ArticleBanner from '@containers/ArticleBanner'
-import RepoContent from '@containers/RepoContent'
+import JobContent from '@containers/JobContent'
 import Footer from '@containers/Footer'
 import ErrorBox from '@containers/ErrorBox'
 
+import { P } from '@schemas'
 import ErrorPage from '@components/ErrorPage'
 
-import { P } from '@schemas'
-
+// import { AnalysisService, ErrorPage } from '@components'
 // try to fix safari bug
 // see https://github.com/yahoo/react-intl/issues/422
 global.Intl = require('intl')
 
-async function fetchData(props, opt) {
-  const { realname } = R.merge({ realname: true }, opt)
-
-  const token = realname ? getJwtToken(props) : null
+async function fetchData(props) {
+  const token = getJwtToken(props)
   const gqClient = makeGQClient(token)
   const userHasLogin = nilOrEmpty(token) === false
 
+  // schema
   const { thridPath: id } = parseURL(props)
 
+  // query data
   const sessionState = gqClient.request(P.sessionState)
-  const repo = gqClient.request(P.repo, { id })
+  const job = gqClient.request(P.job, { id, userHasLogin })
   const pagedComments = gqClient.request(P.pagedComments, {
     id,
     userHasLogin,
-    thread: R.toUpper(THREAD.POST),
+    thread: R.toUpper(THREAD.JOB),
     filter: { page: 1, size: PAGE_SIZE.D, sort: TYPE.ASC_INSERTED },
   })
   const subscribedCommunities = gqClient.request(P.subscribedCommunities, {
@@ -62,31 +64,28 @@ async function fetchData(props, opt) {
 
   return {
     ...(await sessionState),
-    ...(await repo),
+    ...(await job),
     ...(await pagedComments),
     ...(await subscribedCommunities),
   }
 }
 
-export default class RepoPage extends React.Component {
+export default class JobPage extends React.Component {
   static async getInitialProps(props) {
-    const { mainPath, subPath } = parseURL(props)
     let resp
+    const { communityPath, threadPath } = parseURL(props)
+
     try {
       resp = await fetchData(props)
     } catch ({ response: { errors } }) {
       if (ssrAmbulance.hasLoginError(errors)) {
         resp = await fetchData(props, { realname: false })
       } else {
-        return { statusCode: 404, target: subPath }
+        return { statusCode: 404, target: threadPath }
       }
     }
 
-    const { sessionState, repo, pagedComments, subscribedCommunities } = resp
-
-    if (!R.contains(mainPath, R.pluck('raw', repo.communities))) {
-      return { statusCode: 404, target: subPath }
-    }
+    const { sessionState, pagedComments, subscribedCommunities, job } = resp
 
     return {
       langSetup: {},
@@ -98,11 +97,16 @@ export default class RepoPage extends React.Component {
         isValidSession: sessionState.isValid,
         userSubscribedCommunities: subscribedCommunities,
       },
-      route: { mainPath, subPath: ROUTE.REPO },
+      route: {
+        communityPath,
+        mainPath: communityPath,
+        threadPath: ROUTE.JOB,
+        subPath: ROUTE.JOB,
+      },
       viewing: {
-        repo,
-        activeThread: THREAD.REPO,
-        community: repo.origialCommunity,
+        job,
+        activeThread: THREAD.JOB,
+        community: job.origialCommunity,
       },
       comments: { pagedComments },
     }
@@ -120,7 +124,7 @@ export default class RepoPage extends React.Component {
   render() {
     const { statusCode, target } = this.props
     const {
-      viewing: { repo },
+      viewing: { job },
       route,
     } = this.props
     const { mainPath } = route
@@ -130,16 +134,16 @@ export default class RepoPage extends React.Component {
         <AnalysisService>
           <ThemeWrapper>
             {statusCode ? (
-              <ErrorPage errorCode={statusCode} page="post" target={target} />
+              <ErrorPage errorCode={statusCode} page="job" target={target} />
             ) : (
               <React.Fragment>
                 <BlogJsonLd
-                  url={`${SITE_URL}/${mainPath}/repo/${repo.id}`}
-                  title={`${repo.title}`}
-                  datePublished={`${repo.insertedAt}`}
-                  dateModified={`${repo.updatedAt}`}
-                  authorName={`${repo.author.nickname}`}
-                  description={`${repo.desc}`}
+                  url={`${SITE_URL}/${mainPath}/job/${job.id}`}
+                  title={`${job.title}`}
+                  datePublished={`${job.insertedAt}`}
+                  dateModified={`${job.updatedAt}`}
+                  authorName={`${job.author.nickname}`}
+                  description={`${job.title}`}
                   images={[]}
                 />
                 <Route />
@@ -149,12 +153,8 @@ export default class RepoPage extends React.Component {
                   <ErrorBox />
                   <GlobalLayout noSidebar>
                     <Header />
-                    <ArticleBanner
-                      showStar={false}
-                      showWordCount={false}
-                      showLastSync
-                    />
-                    <RepoContent />
+                    <ArticleBanner showStar={false} />
+                    <JobContent />
                     <Footer />
                   </GlobalLayout>
                 </MultiLanguage>
