+++ src/pages/sentry.js
