+++ src/pages/user.js
@@ -1,22 +1,27 @@
+/*
+   this page is for /user/xxx
+ */
 import React from 'react'
 import { Provider } from 'mobx-react'
 import R from 'ramda'
-import { BlogJsonLd } from 'next-seo'
+import { SocialProfileJsonLd } from 'next-seo'
 
-import { PAGE_SIZE, SITE_URL } from '@config'
-import { TYPE, ROUTE, THREAD } from '@constant'
+// eslint-disable-next-line import/named
+import { SITE_URL } from '@config'
+import { ROUTE, USER_THREAD } from '@constant'
 import {
   getJwtToken,
   makeGQClient,
-  parseURL,
+  queryStringToJSON,
   nilOrEmpty,
+  parseURL,
+  pagedFilter,
   ssrAmbulance,
   parseTheme,
 } from '@utils'
 import initRootStore from '@stores/init'
 
 import AnalysisService from '@services/Analysis'
-
 import GlobalLayout from '@containers/GlobalLayout'
 import ThemeWrapper from '@containers/ThemeWrapper'
 import MultiLanguage from '@containers/MultiLanguage'
@@ -24,67 +29,62 @@ import Preview from '@containers/Preview'
 import Doraemon from '@containers/Doraemon'
 import Route from '@containers/Route'
 import Header from '@containers/Header'
-import ArticleBanner from '@containers/ArticleBanner'
-import JobContent from '@containers/JobContent'
+import UserBanner from '@containers/UserBanner'
+import UserContent from '@containers/UserContent'
 import Footer from '@containers/Footer'
 import ErrorBox from '@containers/ErrorBox'
 
 import { P } from '@schemas'
 import ErrorPage from '@components/ErrorPage'
-
 // import { AnalysisService, ErrorPage } from '@components'
+
 // try to fix safari bug
 // see https://github.com/yahoo/react-intl/issues/422
 global.Intl = require('intl')
 
-async function fetchData(props) {
-  const token = getJwtToken(props)
+async function fetchData(props, opt) {
+  const { realname } = R.merge({ realname: true }, opt)
+
+  const token = realname ? getJwtToken(props) : null
   const gqClient = makeGQClient(token)
   const userHasLogin = nilOrEmpty(token) === false
+  const { subPath } = parseURL(props)
+  const login = R.toLower(subPath)
 
-  // schema
-  const { thridPath: id } = parseURL(props)
-
-  // query data
   const sessionState = gqClient.request(P.sessionState)
-  const job = gqClient.request(P.job, { id, userHasLogin })
-  const pagedComments = gqClient.request(P.pagedComments, {
-    id,
-    userHasLogin,
-    thread: R.toUpper(THREAD.JOB),
-    filter: { page: 1, size: PAGE_SIZE.D, sort: TYPE.ASC_INSERTED },
-  })
+  const user = gqClient.request(P.user, { login, userHasLogin })
+
+  const filter = pagedFilter(1)
   const subscribedCommunities = gqClient.request(P.subscribedCommunities, {
-    filter: {
-      page: 1,
-      size: 30,
-    },
+    filter,
   })
 
   return {
     ...(await sessionState),
-    ...(await job),
-    ...(await pagedComments),
+    ...(await user),
     ...(await subscribedCommunities),
   }
 }
 
-export default class JobPage extends React.Component {
+export default class UserPage extends React.Component {
   static async getInitialProps(props) {
-    let resp
-    const { communityPath, threadPath } = parseURL(props)
+    const { asPath } = props
 
+    const query = queryStringToJSON(asPath)
+    const { subPath } = parseURL(props)
+
+    let resp
     try {
       resp = await fetchData(props)
     } catch ({ response: { errors } }) {
       if (ssrAmbulance.hasLoginError(errors)) {
         resp = await fetchData(props, { realname: false })
       } else {
-        return { statusCode: 404, target: threadPath }
+        return { statusCode: 404, target: subPath }
       }
     }
 
-    const { sessionState, pagedComments, subscribedCommunities, job } = resp
+    const { sessionState, user, subscribedCommunities } = resp
 
     return {
       langSetup: {},
@@ -96,23 +96,15 @@ export default class JobPage extends React.Component {
         isValidSession: sessionState.isValid,
         userSubscribedCommunities: subscribedCommunities,
       },
-      route: {
-        communityPath,
-        mainPath: communityPath,
-        threadPath: ROUTE.JOB,
-        subPath: ROUTE.JOB,
-      },
-      viewing: {
-        job,
-        activeThread: THREAD.JOB,
-        community: job.origialCommunity,
-      },
-      comments: { pagedComments },
+      route: { mainPath: ROUTE.USER, subPath: user.id, query },
+      userContent: { activeThread: query.tab || USER_THREAD.PUBLISH },
+      viewing: { user },
     }
   }
 
   constructor(props) {
     super(props)
+
     const store = props.statusCode
       ? initRootStore({ langSetup: {} })
       : initRootStore({ ...props })
@@ -123,27 +115,22 @@ export default class JobPage extends React.Component {
   render() {
     const { statusCode, target } = this.props
     const {
-      viewing: { job },
-      route,
+      viewing: { user },
     } = this.props
-    const { mainPath } = route
 
     return (
       <Provider store={this.store}>
         <AnalysisService>
           <ThemeWrapper>
             {statusCode ? (
-              <ErrorPage errorCode={statusCode} page="job" target={target} />
+              <ErrorPage errorCode={statusCode} page="user" target={target} />
             ) : (
               <React.Fragment>
-                <BlogJsonLd
-                  url={`${SITE_URL}/${mainPath}/job/${job.id}`}
-                  title={`${job.title}`}
-                  datePublished={`${job.insertedAt}`}
-                  dateModified={`${job.updatedAt}`}
-                  authorName={`${job.author.nickname}`}
-                  description={`${job.title}`}
-                  images={[]}
+                <SocialProfileJsonLd
+                  type="Person"
+                  name={`${user.nickname}`}
+                  url={`${SITE_URL}/user/${user.login}`}
+                  sameAs={[]}
                 />
                 <Route />
                 <MultiLanguage>
@@ -152,8 +139,8 @@ export default class JobPage extends React.Component {
                   <ErrorBox />
                   <GlobalLayout noSidebar>
                     <Header />
-                    <ArticleBanner showStar={false} />
-                    <JobContent />
+                    <UserBanner />
+                    <UserContent />
                     <Footer />
                   </GlobalLayout>
                 </MultiLanguage>
