+++ src/stores/index.js
@@ -14,9 +14,7 @@ export { default as PreviewStore } from '@containers/Preview/store'
 export { default as DoraemonStore } from '@containers/Doraemon/store'
 export { default as HeaderStore } from '@containers/Header/store'
 export { default as ContentStore } from '@containers/Content/store'
-export {
-  default as UpgradePackgesStore,
-} from '@containers/UpgradePackges/store'
+export { default as UpgradePackgesStore } from '@containers/UpgradePackges/store'
 export { default as MailBoxStore } from '@containers/MailBox/store'
 export { default as AvatarAdderStore } from '@containers/AvatarAdder/store'
 
@@ -25,36 +23,24 @@ export { default as InformerStore } from '@containers/Informer/store'
 export { default as GirlVerifierStore } from '@containers/GirlVerifier/store'
 export { default as CashierStore } from '@containers/Cashier/store'
 
-export {
-  default as ArticleAuthorCardStore,
-} from '@containers/ArticleAuthorCard/store'
+export { default as ArticleAuthorCardStore } from '@containers/ArticleAuthorCard/store'
 //
 export { default as FavoritesCatsStore } from '@containers/FavoritesCats/store'
 
 // pages banners store
 export { default as BannerStore } from '@containers/Banner/store'
 
-export {
-  default as CommunityBannerStore,
-} from '@containers/CommunityBanner/store'
-export {
-  default as CommunitiesBannerStore,
-} from '@containers/CommunitiesBanner/store'
+export { default as CommunityBannerStore } from '@containers/CommunityBanner/store'
+export { default as CommunitiesBannerStore } from '@containers/CommunitiesBanner/store'
 export { default as UserBannerStore } from '@containers/UserBanner/store'
 
 export { default as ArticleBannerStore } from '@containers/ArticleBanner/store'
 
 // contents store
-export {
-  default as CommunityContentStore,
-} from '@containers/CommunityContent/store'
+export { default as CommunityContentStore } from '@containers/CommunityContent/store'
 
-export {
-  default as CommunitiesContentStore,
-} from '@containers/CommunitiesContent/store'
-export {
-  default as CheatSheetContentStore,
-} from '@containers/CheatSheetContent/store'
+export { default as CommunitiesContentStore } from '@containers/CommunitiesContent/store'
+export { default as CheatSheetContentStore } from '@containers/CheatSheetContent/store'
 export { default as PostContentStore } from '@containers/PostContent/store'
 export { default as JobContentStore } from '@containers/JobContent/store'
 export { default as VideoContentStore } from '@containers/VideoContent/store'
@@ -71,9 +57,7 @@ export { default as ReposThreadStore } from '@containers/ReposThread/store'
 export { default as WikiThreadStore } from '@containers/WikiThread/store'
 export { default as JobsThreadStore } from '@containers/JobsThread/store'
 export { default as UsersThreadStore } from '@containers/UsersThread/store'
-export {
-  default as CheatsheetThreadStore,
-} from '@containers/CheatsheetThread/store'
+export { default as CheatsheetThreadStore } from '@containers/CheatsheetThread/store'
 
 export { default as TagsBarStore } from '@containers/TagsBar/store'
 
@@ -86,9 +70,7 @@ export { default as RepoEditorStore } from '@containers/RepoEditor/store'
 export { default as CommentsStore } from '@containers/Comments/store'
 export { default as AccountEditorStore } from '@containers/AccountEditor/store'
 export { default as LabelerStore } from '@containers/Labeler/store'
-export {
-  default as CommunitySetterStore,
-} from '@containers/CommunitySetter/store'
+export { default as CommunitySetterStore } from '@containers/CommunitySetter/store'
 
 // viewers store
 export { default as PostViewerStore } from '@containers/PostViewer/store'
@@ -97,20 +79,14 @@ export { default as VideoViewerStore } from '@containers/VideoViewer/store'
 export { default as RepoViewerStore } from '@containers/RepoViewer/store'
 export { default as MailsViewerStore } from '@containers/MailsViewer/store'
 
-export {
-  default as ArticleViewerHeader,
-} from '@containers/ArticleViewerHeader/store'
-export {
-  default as ArticleBodyHeaderStore,
-} from '@containers/ArticleBodyHeader/store'
+export { default as ArticleViewerHeader } from '@containers/ArticleViewerHeader/store'
+export { default as ArticleBodyHeaderStore } from '@containers/ArticleBodyHeader/store'
 
 export { default as AccountViewerStore } from '@containers/AccountViewer/store'
 
 // user page
 export { default as UserPublishedStore } from '@containers/UserPublished/store'
-export {
-  default as UserPublishedCommentsStore,
-} from '@containers/UserPublishedComments/store'
+export { default as UserPublishedCommentsStore } from '@containers/UserPublishedComments/store'
 export { default as UserStaredStore } from '@containers/UserStared/store'
 export { default as UserBillingStore } from '@containers/UserBilling/store'
 export { default as UserSettingsStore } from '@containers/UserSettings/store'
