+++ utils/validator.js
@@ -1,19 +1,10 @@
 import R from 'ramda'
 
-export const notEmpty = R.compose(
-  R.not,
-  R.isEmpty
-)
-export const isEmptyValue = R.compose(
-  R.isEmpty,
-  R.trim
-)
+export const notEmpty = R.compose(R.not, R.isEmpty)
+export const isEmptyValue = R.compose(R.isEmpty, R.trim)
 export const nilOrEmpty = R.either(R.isNil, R.isEmpty)
 
-export const hasValue = R.compose(
-  R.not,
-  nilOrEmpty
-)
+export const hasValue = R.compose(R.not, nilOrEmpty)
 
 export const isObject = value => {
   const type = typeof value
@@ -27,15 +18,9 @@ export const isString = value => {
   return false
 }
 
-const notNil = R.compose(
-  R.not,
-  R.isNil
-)
+const notNil = R.compose(R.not, R.isNil)
 
-const validObjects = R.compose(
-  R.pickBy(notNil),
-  R.pickBy(isObject)
-)
+const validObjects = R.compose(R.pickBy(notNil), R.pickBy(isObject))
 
 const emptyArray = obj => Array.isArray(obj) && obj.length === 0
 
@@ -57,14 +42,8 @@ export const cast = (fields, source) => {
   return R.merge(validValues(casted), validObjects(casted))
 }
 
-const keyOf = R.compose(
-  R.head,
-  R.keys
-)
-const valueOf = R.compose(
-  R.head,
-  R.values
-)
+const keyOf = R.compose(R.head, R.keys)
+const valueOf = R.compose(R.head, R.values)
 
 export const changeset = source => ({
   exsit: (obj, cb, opt = { skip: false }) => {
